1\MetaServer.java,2\TestMetaClient.java

+++ paraflow-metaserver/src/main/java/cn/edu/ruc/iir/paraflow/metaserver/client/MetaClient.java
@@ -49,20 +49,21 @@ public class MetaClient
         MetaProto.CreateUserParam createUser = MetaProto.CreateUserParam.newBuilder().setUserName(userName).setCreateTime(createTime).setLastVisitTime(lastVisitTime).build();
         MetaProto.StatusType status;
         try {
+            System.out.println("1status = metaBlockingStub.createUser(createUser)");
             status = metaBlockingStub.createUser(createUser);
+            System.out.println("2status = metaBlockingStub.createUser(createUser)");
         }
         catch (StatusRuntimeException e) {
             logger.log(Level.WARNING, "RPC failed: {0}", e.getStatus());
             status = MetaProto.StatusType.newBuilder().build();
             return status;
         }
-        logger.info("Rename column status is : " + status.getStatus());
+        logger.info("Create user status is : " + status.getStatus());
         return status;
     }
 
     public MetaProto.StatusType createDatabase(String dbName, String locationUrl, String userName)
     {
-        //MetaProto.UserParam user = MetaProto.UserParam.newBuilder().setUserName("Alice").setUserPass("123456").setRoleName("admin").setCreationTime(20170807).setLastVisitTime(20170807).build();
         MetaProto.DbParam database = MetaProto.DbParam.newBuilder().setDbName(dbName).setLocationUrl(locationUrl).setUserName(userName).build();
         MetaProto.StatusType status;
         try {
@@ -77,28 +78,27 @@ public class MetaClient
         return status;
     }
 
-    public MetaProto.StatusType createTable(String dbName, String tblName, int tblType, int createTime, int lastAccessTime, String locationUrl, int storageFormatId, int fiberColId, int fiberFuncId, ArrayList<String> columnName, ArrayList<String> columnType, ArrayList<String> datatype)
+    public MetaProto.StatusType createTable(String dbName, String tblName, int tblType, String userName, int createTime, int lastAccessTime, String locationUrl, int storageFormatId, int fiberColId, int fiberFuncId, ArrayList<String> columnName, ArrayList<String> columnType, ArrayList<String> dataType)
     {
-//        MetaProto.UserParam user = MetaProto.UserParam.newBuilder().setUserName("Alice").setUserPass("123456").setRoleName("admin").setCreationTime(20170807).setLastVisitTime(20170807).build();
-//        MetaProto.DbParam database = MetaProto.DbParam.newBuilder().setName("default").setLocationUri("hdfs:/127.0.0.1:9000/warehouse/default").setUser(user).build();
         int number = columnName.size();
         ArrayList<MetaProto.ColParam> columns = new ArrayList<>();
         for (int i = 0; i < number; i++) {
-            MetaProto.ColParam column = MetaProto.ColParam.newBuilder().setColIndex(i).setTblName(tblName).setColName(columnName.get(i)).setColType(columnType.get(i)).setDataType(datatype.get(i)).build();
+            MetaProto.ColParam column = MetaProto.ColParam.newBuilder().setColIndex(i).setDbName(dbName).setTblName(tblName).setColName(columnName.get(i)).setColType(columnType.get(i)).setDataType(dataType.get(i)).build();
             columns.add(column);
         }
         MetaProto.ColListType colList = MetaProto.ColListType.newBuilder().addAllColumn(columns).build();
-        MetaProto.TblParam table = MetaProto.TblParam.newBuilder().setDbName(dbName).setTblName(tblName).setTblType(tblType).setCreateTime(createTime).setLastAccessTime(lastAccessTime).setLocationUrl(locationUrl).setStorageFormatId(storageFormatId).setFiberColId(fiberColId).setFiberFuncId(fiberFuncId).build();
+        MetaProto.TblParam table = MetaProto.TblParam.newBuilder().setDbName(dbName).setTblName(tblName).setUserName(userName).setTblType(tblType).setCreateTime(createTime).setLastAccessTime(lastAccessTime).setLocationUrl(locationUrl).setStorageFormatId(storageFormatId).setFiberColId(fiberColId).setFiberFuncId(fiberFuncId).build();
         MetaProto.StatusType statusColumn;
         MetaProto.StatusType statusTable;
         try {
-            statusColumn = metaBlockingStub.createColumn(colList);
-            MetaProto.StatusType statusOK = MetaProto.StatusType.newBuilder().setStatus(MetaProto.StatusType.State.OK).build();
-            if (statusColumn == statusOK) {
-                statusTable = metaBlockingStub.createTable(table);
+            statusTable = metaBlockingStub.createTable(table);
+            //System.out.println("create table status is" + statusTable.getStatus());
+            //MetaProto.StatusType statusOK = MetaProto.StatusType.newBuilder().setStatus(MetaProto.StatusType.State.OK).build();
+            if (statusTable.getStatus() == MetaProto.StatusType.State.OK) {
+                statusColumn = metaBlockingStub.createColumn(colList);
             }
             else {
-                MetaProto.StatusType statusError = MetaProto.StatusType.newBuilder().setStatus(MetaProto.StatusType.State.CREAT_COLUMN_ERROR).build();
+                MetaProto.StatusType statusError = MetaProto.StatusType.newBuilder().setStatus(MetaProto.StatusType.State.CREATE_TABLE_ERROR).build();
                 return statusError;
             }
         }
@@ -107,8 +107,8 @@ public class MetaClient
             MetaProto.StatusType statusError = MetaProto.StatusType.newBuilder().setStatus(MetaProto.StatusType.State.CREAT_COLUMN_ERROR).build();
             return statusError;
         }
-        logger.info("Create table status is : " + statusTable.getStatus());
-        return statusTable;
+        logger.info("Create table status is : " + statusColumn.getStatus());
+        return statusColumn;
     }
 
     public MetaProto.StringListType listDatabases()
@@ -252,33 +252,63 @@ public class MetaClient
         MetaProto.DbNameParam databaseName = MetaProto.DbNameParam.newBuilder().setDatabase(dbName).build();
         MetaProto.TblNameParam tableName = MetaProto.TblNameParam.newBuilder().setTable(tblName).build();
         MetaProto.DbTblParam databaseTable = MetaProto.DbTblParam.newBuilder().setDatabase(databaseName).setTable(tableName).build();
-        MetaProto.StatusType status;
+        MetaProto.StatusType statusColumn;
+        MetaProto.StatusType statusTable;
         try {
-            status = metaBlockingStub.deleteTable(databaseTable);
+            statusColumn = metaBlockingStub.deleteTblColumn(databaseTable);
+            //System.out.println("create table status is" + statusTable.getStatus());
+            //MetaProto.StatusType statusOK = MetaProto.StatusType.newBuilder().setStatus(MetaProto.StatusType.State.OK).build();
+            System.out.println("delete column state is : " + statusColumn.getStatus());
+            if (statusColumn.getStatus() == MetaProto.StatusType.State.OK) {
+                statusTable = metaBlockingStub.deleteTable(databaseTable);
+            }
+            else {
+                MetaProto.StatusType statusError = MetaProto.StatusType.newBuilder().setStatus(MetaProto.StatusType.State.DELETE_COLUMN_ERROR).build();
+                return statusError;
+            }
         }
         catch (StatusRuntimeException e) {
             logger.log(Level.WARNING, "RPC failed: {0}", e.getStatus());
-            status = MetaProto.StatusType.newBuilder().build();
-            return status;
+            statusTable = MetaProto.StatusType.newBuilder().build();
+            return statusTable;
         }
-        logger.info("Delete table status is : " + status.getStatus());
-        return status;
+        logger.info("Delete table status is : " + statusTable.getStatus());
+        return statusTable;
     }
 
     public MetaProto.StatusType deleteDatabase(String dbName)
     {
         MetaProto.DbNameParam databaseName = MetaProto.DbNameParam.newBuilder().setDatabase(dbName).build();
-        MetaProto.StatusType status;
+        MetaProto.StatusType statusDb;
+        MetaProto.StatusType statusCol;
+        MetaProto.StatusType statusTbl;
         try {
-            status = metaBlockingStub.deleteDatabase(databaseName);
+            statusCol = metaBlockingStub.deleteDbColumn(databaseName);
+            //System.out.println("create table status is" + statusTable.getStatus());
+            //MetaProto.StatusType statusOK = MetaProto.StatusType.newBuilder().setStatus(MetaProto.StatusType.State.OK).build();
+            //System.out.println("delete column state is : " + statusColumn.getStatus());
+            if (statusCol.getStatus() == MetaProto.StatusType.State.OK) {
+                statusTbl = metaBlockingStub.deleteDbTable(databaseName);
+                if (statusTbl.getStatus() == MetaProto.StatusType.State.OK) {
+                    statusDb = metaBlockingStub.deleteDatabase(databaseName);
+            }
+            else {
+                    MetaProto.StatusType statusError = MetaProto.StatusType.newBuilder().setStatus(MetaProto.StatusType.State.DELETE_TABLE_ERROR).build();
+                    return statusError;
+                }
+            }
+            else {
+                MetaProto.StatusType statusError = MetaProto.StatusType.newBuilder().setStatus(MetaProto.StatusType.State.DELETE_COLUMN_ERROR).build();
+                return statusError;
+            }
         }
         catch (StatusRuntimeException e) {
             logger.log(Level.WARNING, "RPC failed: {0}", e.getStatus());
-            status = MetaProto.StatusType.newBuilder().build();
-            return status;
+            statusDb = MetaProto.StatusType.newBuilder().build();
+            return statusDb;
         }
-        logger.info("Delete database status is : " + status.getStatus());
-        return status;
+        logger.info("Delete database status is : " + statusDb.getStatus());
+        return statusDb;
     }
 
     public MetaProto.StatusType createDbParam(String dbName, String paramKey, String paramValue)
@@ -293,13 +323,13 @@ public class MetaClient
             status = MetaProto.StatusType.newBuilder().build();
             return status;
         }
-        logger.info("Rename column status is : " + status.getStatus());
+        logger.info("Create database param status is : " + status.getStatus());
         return status;
     }
 
-    public MetaProto.StatusType createTblParam(String tblName, String paramKey, String paramValue)
+    public MetaProto.StatusType createTblParam(String dbName, String tblName, String paramKey, String paramValue)
     {
-        MetaProto.CreateTblParamParam createTblParam = MetaProto.CreateTblParamParam.newBuilder().setTblName(tblName).setParamKey(paramKey).setParamValue(paramValue).build();
+        MetaProto.CreateTblParamParam createTblParam = MetaProto.CreateTblParamParam.newBuilder().setDbName(dbName).setTblName(tblName).setParamKey(paramKey).setParamValue(paramValue).build();
         MetaProto.StatusType status;
         try {
             status = metaBlockingStub.createTblParam(createTblParam);
@@ -309,13 +339,13 @@ public class MetaClient
             status = MetaProto.StatusType.newBuilder().build();
             return status;
         }
-        logger.info("Rename column status is : " + status.getStatus());
+        logger.info("Create table param status is : " + status.getStatus());
         return status;
     }
 
-    public MetaProto.StatusType createTblPriv(String tblName, int privType, int grantTime)
+    public MetaProto.StatusType createTblPriv(String dbName, String tblName, String userName, int privType, int grantTime)
     {
-        MetaProto.CreateTblPrivParam createTblPriv = MetaProto.CreateTblPrivParam.newBuilder().setTblName(tblName).setPrivType(privType).setGrantTime(grantTime).build();
+        MetaProto.CreateTblPrivParam createTblPriv = MetaProto.CreateTblPrivParam.newBuilder().setDbName(dbName).setTblName(tblName).setUserName(userName).setPrivType(privType).setGrantTime(grantTime).build();
         MetaProto.StatusType status;
         try {
             status = metaBlockingStub.createTblPriv(createTblPriv);
@@ -325,7 +355,7 @@ public class MetaClient
             status = MetaProto.StatusType.newBuilder().build();
             return status;
         }
-        logger.info("Rename column status is : " + status.getStatus());
+        logger.info("Create tblpriv status is : " + status.getStatus());
         return status;
     }
 
@@ -341,7 +371,7 @@ public class MetaClient
             status = MetaProto.StatusType.newBuilder().build();
             return status;
         }
-        logger.info("Rename column status is : " + status.getStatus());
+        logger.info("Create storage format status is : " + status.getStatus());
         return status;
     }
 
@@ -357,7 +387,7 @@ public class MetaClient
             status = MetaProto.StatusType.newBuilder().build();
             return status;
         }
-        logger.info("Rename column status is : " + status.getStatus());
+        logger.info("Create fiber function status is : " + status.getStatus());
         return status;
     }
 
@@ -376,7 +406,7 @@ public class MetaClient
             status = MetaProto.StatusType.newBuilder().build();
             return status;
         }
-        logger.info("Add block index status is : " + status.getStatus());
+        logger.info("Create block index status is : " + status.getStatus());
         return status;
     }
